requirements.txt
attrs
pandas
pandas_bokeh
yfinance
python-exchangeratesapi